refactor(storeprovider): refactored the store provider location directory
