--- app/StoreProvider.tsx
@@ -1,19 +1,0 @@
-"use client";
-import { Provider } from "react-redux";
-import {  persistor, store } from "../lib/store";
-import { PersistGate } from "redux-persist/integration/react";
-
-export default function StoreProvider({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-
-  return (
-    <Provider store={store}>
-      <PersistGate loading={null} persistor={persistor}>
-        {children}
-      </PersistGate>
-    </Provider>
-  );
-}
