+++ app/layout.tsx
@@ -3,8 +3,7 @@ import { Inter } from "next/font/google";
 import "./globals.css";
 import { ThemeProvider } from "@/components/providers/theme-provider";
 import { Toaster } from "@/components/ui/sonner";
-import { StoreProvider } from "../components/providers/store-provider";
-// import StoreProvider from "./StoreProvider";
+import StoreProvider from "@/components/providers/store-provider";
 
 const inter = Inter({
   subsets: ["latin"],
