+++ components/providers/store-provider.tsx
@@ -1,10 +1,19 @@
 "use client";
-
-import { makeStore } from "@/lib/store";
+import { persistor, store } from "@/lib/store";
 import { Provider } from "react-redux";
+import { PersistGate } from "redux-persist/integration/react";
 
-const store = makeStore();
+export default function StoreProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
 
-export function StoreProvider({ children }: { children: React.ReactNode }) {
-  return <Provider store={store}>{children}</Provider>;
+  return (
+    <Provider store={store}>
+      <PersistGate loading={null} persistor={persistor}>
+        {children}
+      </PersistGate>
+    </Provider>
+  );
 }
